requirements.txt
# -----------------------------
# Web Framework
# -----------------------------
fastapi>=0.110.0
uvicorn>=0.29.0

# -----------------------------
# CrewAI
# -----------------------------
crewai==0.203.1
crewai-tools==0.76.0

# -----------------------------
# LangChain
# -----------------------------
langchain==0.3.27
langchain-community==0.3.31
langchain-openai==0.3.35

# -----------------------------
# OpenAI SDK
# -----------------------------
openai>=1.104.2,<3.0.0

# -----------------------------
# Data
# -----------------------------
pandas>=2.2.0
numpy>=1.26.0

# -----------------------------
# Utilities (DO NOT PIN)
# -----------------------------
requests>=2.32.5
PyYAML>=6.0.1
python-dotenv>=1.1.1
